requirements.txt
av==12.3.0
beautifulsoup4==4.12.3
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.3.1
deep-translator==1.11.4
faster-whisper==1.0.3
ffmpeg-python==0.2.0
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
future==1.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.24.6
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
mpmath==1.3.0
numpy==1.24.4
onnxruntime==1.19.0
packaging==24.1
pandas==2.0.3
protobuf==5.27.3
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.2
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
